refactor: move AdUpdateStrategy to ad_model and extract price_reduction module
- Add AdUpdateStrategy enum to src/kleinanzeigen_bot/model/ad_model.py
- Extract price reduction logic into src/kleinanzeigen_bot/price_reduction.py
  (_repost_delay_state, _day_delay_state, PriceReductionDecision,
   evaluate_auto_price_reduction, _log_auto_price_reduction_preview,
   apply_auto_price_reduction)
- Reduce __init__.py from 4245 to 3778 lines (467 lines removed)
- Preserve root-package re-exports for backward compatibility
- Update translations.de.yaml for new module locations
- Add root re-export verification test
